+++ src/app/MainPageFiles/jsx/jsx-Footer/MainPageFooter.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import styles from "../../mainPageStyles/MainPageFooter-styles.module.css";
+import styles from '../../styles/MainPageFooter-styles.module.css';
 import MainPageFooterList from "./MainPageFooterList";
-import MainPageFooterAddress from "./MainPageFooterAddress";
 function MainPageFooter(props) {
 	return (
 		<div className={styles.footer}>
