+++ src/app/MainPageFiles/jsx/jsx-Header/MainPageHeaderBasket.jsx
@@ -1,17 +1,18 @@
 import React from "react";
 import Link from "next/link";
 import Image from "next/image";
-import styles from "../../mainPageStyles/MainPageHeader-styles.module.css";
+import styles from '../../styles/MainPageHeader-styles.module.css';
 function MainPageHeaderBasket(props) {
   
   return (
     <div id={styles.mainPageHeaderLogoContainer}>
-      <img
+      <Image
         className={styles.mainPageHeaderBasketPhoto}
         src={props.basketPhotoUrl}
         alt="logo"
         width={props.basketPhotoWidth}
         height={props.basketPhotoHeight}
+        priority={true}
       />
       <Link href="/">Кошик</Link>
     </div>
