+++ src/app/MainPageFiles/jsx/jsx-Header/MainPageHeaderMenu.jsx
@@ -1,16 +1,20 @@
 import React from "react";
 import Link from "next/link";
 import Image from "next/image";
-import styles from "../../mainPageStyles/MainPageHeader-styles.module.css";
+import styles from '../../styles/MainPageHeader-styles.module.css';
 function MainPageHeaderMenu(props) {
   return (
     <div id={styles.mainPageHeaderLogoContainer}>
-      <img
+      <Image
         id={styles.mainPageHeaderMenu}
         src={props.photoUrl}
         alt="logo"
-        width={props.photoWidth}
-        height={props.photoHeight}
+        style={{ 
+          cursor: "pointer"
+         }}
+         width={props.photoWidth}
+         height={props.photoHeight}
+        priority
       />
       <Link href="../MainPage/index.html">Меню</Link>
     </div>
